slsReceiver bug fix affecting Jungfrau receivers as root (optlen size as int), but setting buffer size is uint32_t size. setting it to uint64_t

+++ slsDetectorSoftware/multiSlsDetector/multiSlsDetector.cpp
@@ -2002,7 +2002,7 @@ int multiSlsDetector::setDetectorMode(detectorModeType value, int detPos) {
     return getDetectorModeType(result);
 }
 
-int multiSlsDetector::setReceiverUDPSocketBufferSize(int udpsockbufsize, int detPos) {
+uint64_t multiSlsDetector::setReceiverUDPSocketBufferSize(uint64_t udpsockbufsize, int detPos) {
     // single
     if (detPos >= 0) {
         return detectors[detPos]->setReceiverUDPSocketBufferSize(udpsockbufsize);
@@ -2013,7 +2013,7 @@ int multiSlsDetector::setReceiverUDPSocketBufferSize(int udpsockbufsize, int det
     return sls::minusOneIfDifferent(r);
 }
 
-int multiSlsDetector::getReceiverUDPSocketBufferSize(int detPos) {
+uint64_t multiSlsDetector::getReceiverUDPSocketBufferSize(int detPos) {
     // single
     if (detPos >= 0) {
         return detectors[detPos]->getReceiverUDPSocketBufferSize();
@@ -2024,7 +2024,7 @@ int multiSlsDetector::getReceiverUDPSocketBufferSize(int detPos) {
     return sls::minusOneIfDifferent(r);
 }
 
-int multiSlsDetector::getReceiverRealUDPSocketBufferSize(int detPos) {
+uint64_t multiSlsDetector::getReceiverRealUDPSocketBufferSize(int detPos) {
     // single
     if (detPos >= 0) {
         return detectors[detPos]->getReceiverRealUDPSocketBufferSize();
