+++ slsDetectorSoftware/slsDetector/slsDetector.h
@@ -999,19 +999,19 @@ public:
 	 * @param udpsockbufsize additional json header
 	 * @returns receiver udp socket buffer size
 	 */
-	int setReceiverUDPSocketBufferSize(int udpsockbufsize=-1);
+   uint64_t setReceiverUDPSocketBufferSize(uint64_t udpsockbufsize=-1);
 
 	/**
 	 * Returns the receiver UDP socket buffer size\sa sharedSlsDetector
 	 * @returns the receiver UDP socket buffer size
 	 */
-	int getReceiverUDPSocketBufferSize() ;
+   uint64_t getReceiverUDPSocketBufferSize() ;
 
 	/**
 	 * Returns the receiver real UDP socket buffer size\sa sharedSlsDetector
 	 * @returns the receiver real UDP socket buffer size
 	 */
-	int getReceiverRealUDPSocketBufferSize();
+   uint64_t getReceiverRealUDPSocketBufferSize();
 
 	/**
 	 * Execute a digital test (Gotthard, Mythen)
